+++ src/hassette/models/states/automation.py
@@ -1,7 +1,7 @@
 from typing import Literal
 
 from pydantic import Field, field_validator
-from whenever import SystemDateTime
+from whenever import ZonedDateTime
 
 from hassette.utils.date_utils import convert_datetime_str_to_system_tz
 
@@ -11,14 +11,14 @@ from .base import AttributesBase, StringBaseState
 class AutomationState(StringBaseState):
     class Attributes(AttributesBase):
         id: str | None = Field(default=None)
-        last_triggered: SystemDateTime | None = Field(default=None)
+        last_triggered: ZonedDateTime | None = Field(default=None)
         mode: str | None = Field(default=None)
         current: int | float | None = Field(default=None)
         max: int | float | None = Field(default=None)
 
         @field_validator("last_triggered", mode="before")
         @classmethod
-        def parse_last_triggered(cls, value: SystemDateTime | str | None) -> SystemDateTime | None:
+        def parse_last_triggered(cls, value: ZonedDateTime | str | None) -> ZonedDateTime | None:
             return convert_datetime_str_to_system_tz(value)
 
     domain: Literal["automation"]